Clean up some warnings Fix ShaderBuilder destruction Change warning level Delete some commented out things

+++ JetEngine/CInput.h
@@ -137,8 +137,8 @@ public:
 	{
 		//caps.Vibration 
 		XINPUT_VIBRATION vib;
-		vib.wLeftMotorSpeed = left * 65535.0f;
-		vib.wRightMotorSpeed = right * 65535.0f;
+		vib.wLeftMotorSpeed = static_cast<unsigned short>(left * 65535.0f);
+		vib.wRightMotorSpeed = static_cast<unsigned short>(right * 65535.0f);
 		XInputSetState(id, &vib);
 	}
